feat: extracts redux-offline functionality to its own folder
